refactor(geometry): split processors/advanced_face.rs into 7 concern modules
advanced_face.rs (1574 LOC) -> advanced_face/{mod (dispatcher), bspline (NURBS
math), curves (per-edge samplers), edge_loop (topology walk), polyline (SoR
generator profile), surfaces (planar/bspline/cylindrical processors),
revolution}. Split granularly so every resulting file is under ~400 LOC.
Pure code motion; cross-submodule items pub(super) (re-exported items pub(crate)
to satisfy E0364); processors/mod.rs + sibling callers untouched. build/clippy
-D warnings clean; advanced_face + annex_e snapshot tests green.
